move other collection serializers into collection
