+++ java/fury-core/src/main/java/io/fury/resolver/ClassResolver.java
@@ -48,12 +48,12 @@ import io.fury.exception.InsecureException;
 import io.fury.memory.MemoryBuffer;
 import io.fury.serializer.ArraySerializers;
 import io.fury.serializer.BufferSerializers;
-import io.fury.serializer.ChildContainerSerializers;
+import io.fury.serializer.collection.ChildContainerSerializers;
 import io.fury.serializer.CodegenSerializer.LazyInitBeanSerializer;
 import io.fury.serializer.CompatibleSerializer;
 import io.fury.serializer.ExternalizableSerializer;
-import io.fury.serializer.GuavaSerializers;
-import io.fury.serializer.ImmutableCollectionSerializers;
+import io.fury.serializer.collection.GuavaCollectionSerializers;
+import io.fury.serializer.collection.ImmutableCollectionSerializers;
 import io.fury.serializer.JavaSerializer;
 import io.fury.serializer.JdkProxySerializer;
 import io.fury.serializer.LambdaSerializer;
@@ -68,12 +68,12 @@ import io.fury.serializer.SerializerFactory;
 import io.fury.serializer.Serializers;
 import io.fury.serializer.StringSerializer;
 import io.fury.serializer.StructSerializer;
-import io.fury.serializer.SynchronizedSerializers;
+import io.fury.serializer.collection.SynchronizedSerializers;
 import io.fury.serializer.TimeSerializers;
 import io.fury.serializer.UnexistedClassSerializers.UnexistedClassSerializer;
 import io.fury.serializer.UnexistedClassSerializers.UnexistedMetaSharedClass;
 import io.fury.serializer.UnexistedClassSerializers.UnexistedSkipClass;
-import io.fury.serializer.UnmodifiableSerializers;
+import io.fury.serializer.collection.UnmodifiableSerializers;
 import io.fury.serializer.collection.CollectionSerializer;
 import io.fury.serializer.collection.CollectionSerializers;
 import io.fury.serializer.collection.MapSerializer;
@@ -342,7 +342,7 @@ public class ClassResolver {
     UnmodifiableSerializers.registerSerializers(fury);
     ImmutableCollectionSerializers.registerSerializers(fury);
     if (fury.getConfig().registerGuavaTypes()) {
-      GuavaSerializers.registerDefaultSerializers(fury);
+      GuavaCollectionSerializers.registerDefaultSerializers(fury);
     }
     if (metaContextShareEnabled) {
       addDefaultSerializer(
