+++ java/fury-core/src/main/java/io/fury/serializer/collection/ImmutableCollectionSerializers.java
@@ -14,12 +14,10 @@
  * limitations under the License.
  */
 
-package io.fury.serializer;
+package io.fury.serializer.collection;
 
 import io.fury.Fury;
 import io.fury.memory.MemoryBuffer;
-import io.fury.serializer.collection.CollectionSerializer;
-import io.fury.serializer.collection.MapSerializer;
 import io.fury.util.Platform;
 import io.fury.util.unsafe._JDKAccess;
 import java.lang.invoke.MethodHandle;
