+++ java/fury-core/src/main/java/io/fury/serializer/collection/SubListSerializers.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.fury.serializer;
+package io.fury.serializer.collection;
 
 public class SubListSerializers {
   // ArrayListSubListSerializer
