+++ java/fury-core/src/test/java/io/fury/serializer/collection/UnmodifiableSerializersTest.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package io.fury.serializer;
+package io.fury.serializer.collection;
 
-import static io.fury.serializer.UnmodifiableSerializers.createSerializer;
+import static io.fury.serializer.collection.UnmodifiableSerializers.createSerializer;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
 
@@ -26,6 +26,7 @@ import io.fury.FuryTestBase;
 import io.fury.config.Language;
 import io.fury.memory.MemoryBuffer;
 import io.fury.memory.MemoryUtils;
+import io.fury.serializer.Serializer;
 import io.fury.serializer.collection.CollectionSerializersTest;
 import io.fury.test.bean.CollectionFields;
 import io.fury.test.bean.MapFields;
